Update PixelArtDetector.py
Had an issue where images using uncommon Aspect Ratios were getting resized.

Had ChatGPT identify and resolve the issue, and tested this in ComfyUI-- but it's worth a code review with scrutiny. Maybe we don't need PIL libraries, but it was working for me and figured I'd push this upstream.